requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Login
python-dotenv
requests
Flask-Migrate
Flask-WTF 
bcrypt